Removing the old rsbuffer
